Create MQ2 Gas Sensor
MQ2 gas sensor is an electronic sensor used for sensing the concentration of gases
in the air such as LPG, propane, methane, hydrogen, alcohol, smoke and carbon
monoxide.

MQ2 gas sensor is also known as chemi-resistor. It contains a sensing material
whose resistance changes when it comes in contact with the gas. This change in
the value of resistance is used for the detection of gas.