Create Arduino push button
This Arduino sketch is a basic example of using a push button to control an LED. The LED is connected to pin 13 and is turned on when the push button, connected to pin 2, is pressed. When the button is released, the LED turns off. This code continuously checks the state of the button in the loop() function and adjusts the LED accordingly.